move to util dir
